requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data & numerics
numpy>=1.24.0
Pillow>=9.0.0
pandas>=1.5.0

# Visualisation
matplotlib>=3.7.0

# Config & CLI
pyyaml>=6.0

# Logging
tensorboard>=2.13.0

# Testing
pytest>=7.4.0
